opea-comps/opea-services/image2video/requirements.txt
torch
diffusers
transformers
fastapi
uvicorn
safetensors
accelerate
requests
pillow